# back-end/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "List API Juong Job",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "duongchieu7202@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/seeker/job/id/apply": {
            "post": {
                "tags": [
                    "Application"
                ],
                "summary": "Seeker apply job",
                "description": "Return apply",
                "operationId": "applyForJob",
                "requestBody": {
                    "description": "Job post data",
                    "content": {
                        "application/json": {
                            "example": {
                                "seeker_id": 1,
                                "job_id": 3,
                                "name": "Software Developer",
                                "phone": "Mid-level",
                                "cover_letter": "Full-time",
                                "resume_path": "Java, Python"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Job post created successfully",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Application submitted successfully",
                                    "application": {
                                        "seeker_id": 7,
                                        "job_id": "1",
                                        "name": "toi la duong",
                                        "phone": "1231234567",
                                        "cover_letter": "hiiii",
                                        "resume_path": "1693419346.pdf",
                                        "updated_at": "2023-08-30T18:15:46.000000Z",
                                        "created_at": "2023-08-30T18:15:46.000000Z",
                                        "id": 3
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "An error occurred while creating the job post"
                    }
                }
            }
        },
        "/seeker/apply/history": {
            "get": {
                "tags": [
                    "Application"
                ],
                "summary": "Get application history of seeker",
                "description": "Return history apply",
                "operationId": "getApplicationHistory",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
    }
}